+++ .github/workflows/acceptance.yml
@@ -39,12 +39,16 @@ jobs:
       - name: Install hatch
         run: pip install hatch==1.9.4
 
+      - name: Fetch relevant branches
+        run: |
+          git fetch origin $GITHUB_BASE_REF:$GITHUB_BASE_REF
+          git fetch origin $GITHUB_HEAD_REF:$GITHUB_HEAD_REF
+
       - name: Run integration tests
-        # ...
         uses: databrickslabs/sandbox/acceptance@acceptance/v0.2.2
         with:
           vault_uri: ${{ secrets.VAULT_URI }}
-          timeout: 45m
+          timeout: 55m
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           ARM_CLIENT_ID: ${{ secrets.ARM_CLIENT_ID }}
